minor arg parsing allocation change
follow up to my last pr. before this the page_allocator was being used
in main.loadEnv() and parseArgs(). now main_arena is being used for
this.

i removed `const allocator = std.heap.page_allocator` from main as it
was easy to forget that it was page_allocator and misuse it.