cap log file size to Fuzzer.file_cap
this patch chages the way log files are written to. in order to cap
their size and maintain good perf, we write to the files circularly
while always writing an EOF marker afterward. then when a zls crash is
found, the 3 log files are reordered, eliminating the EOF marker so that
the reader doesn't have to scan for it.

notes:
* the stdout.log grows to double Fuzzer.file_cap for some reason
* when zls crashes, also copy staging/markov/principal.zig into
  saved_logs
* remove duplicate logic from Fuzzer.create by using Fuzzer.reset
* log a heartbeat every 1000 ticks to detect zls stalls
* add script/killzls.sh to quickly kill zls on linux
* add Fuzzer.log_buf - used while reordering logs
* rename
  * Fuzzer.std{in,out,err} => std..._file
  * Fuzzer.read_buf => stdout_buf
  * Fuzzer.write_buf => stdin_buf