cap log file size to Fuzzer.LogFile.file_cap
this patch chages the way log files are written to. in order to cap
their size, we rotate the log files when they grow beyond the cap.
then when a zls crash is found the rotated log files are merged in order
and renamed if necessary before moving into the saved_logs folder.

notes:
* add Fuzzer.LogFile which represents a rotating log file
* when zls crashes, also copy staging/markov/principal.zig into
  saved_logs
* remove duplicate logic from Fuzzer.create by using Fuzzer.reset
* log a heartbeat every 1000 ticks to detect zls stalls
* add script/killzls.sh to quickly kill zls on linux
* rename Fuzzer.std{in,out,err} => std..._file
* replace Fuzzer.{read,write}_buf with Fuzzer.buf